Tier 3a: EIPの32bit化 — 最後の16bit配線を外す (#28)
* nasmのリストファイルを追い出す (デバッグ中の生成物がPR #26に紛れた)

* EIPを32bitにする — 最後の16bit配線を外す

ip: u16 はコアに残った最後の16bit配線だった。64KBを超えるコード
(Linuxカーネルは1MB超) はこれが外れないと歩けない。

## 折り返しの持ち主を1箇所にする

widen して終わりではなく、「16bitコードでは64Kで折り返す」という8086の
挙動を**どこが知っているか**を決めるのが本体だった。散らばらせず、
Cpu::ip_mask() / set_ip() / advance_ip() の3つだけが幅を知る。
幅はモードとCSのDビットから導く — regs を最初から u32 で持っていたのと
同じ判断を IP にも適用した形。

- フェッチ (最熱経路) は advance_ip(1)。ANDが1個増えるだけ
- 分岐族 (jmp/call/ret/jcc/loop/grp5) は着地を set_ip に統一。
  ついでに**オペランドサイズ対応** (rel32 / pop32) まで済ませた —
  型を直すだけなら二度手間になる場所
- far call/retf (9A/CB/CA) はリアルモード16bit形のまま。保護モードの
  far call は特権チェックと絡むので、リングを作るときに一緒にやる

## 道連れ

- 割り込み統計 (int_first/int_recent/first_fault) のIPを u32 に
- debug (Step.ip / Stop::*.at / before_exec) を u32 に
- スナップショット v4
- boot.rs の診断にあった sel<<4 直書きの残党2箇所も cpu.lin() に統一

## 検証

- 新テスト2件: **EIPが64Kを越えて歩ける** (0x2_0000 のコードを実行。
  これまでの全テストは偶然64K以下で走っていた) / **リアルモードのIPは
  64Kで折り返す** (8086の環)
- 103テスト全通過、clippy 0
- **ELKSがまたビット同一 (45,669,277命令)** — フェッチ経路を触っても
  リアルモードの挙動はゼロ変化。FreeDOS も dir まで無傷

* cosim を EIP 32bit化に追随させる

CPU照合 (CI) だけが捕まえたビルド漏れ。cosim は default-members の外に
居るので、手元の cargo build --all-targets も clippy --workspace
--exclude rustx86-cosim も、全部この crate を素通りしていた。
Unicorn のCビルドが重いから除外している以上、ここを見張れるのは
paths で発火する CPU照合 ワークフローだけ — 設計どおりの網に掛かった。

State::ip は u16 のまま、渡す側を切り詰める。MODE_16 の突き合わせでは
IPは16bitが正しい。MODE_32 対応のときに器ごと広げる。

手元でも cargo test -p rustx86-cosim を実際に回した: 15件全通過
(DAAの総当たり含む、112秒)。

+++ core/examples/dbg.rs
@@ -277,7 +277,7 @@ fn why(m: &Machine, s: &Stop) -> String {
 
 fn show_where(m: &Machine) {
     let cpu = &m.cpu;
-    let lin = cpu.lin(rustx86_core::cpu::CS, cpu.ip as u32);
+    let lin = cpu.lin(rustx86_core::cpu::CS, cpu.ip);
     let mut b = String::new();
     for i in 0..8 {
         b.push_str(&format!("{:02x} ", m.read8(lin.wrapping_add(i))));
